RE 2585 Refacted Metis UI. Separated the MetisLandingPage in base class and subclasses
