+++ metis-ui/src/main/java/eu/europeana/metis/common/MetisPage.java
@@ -1,7 +1,7 @@
 package eu.europeana.metis.common;
 
 import de.flapdoodle.embed.process.collections.Collections;
-import eu.europeana.metis.page.PageView;
+import eu.europeana.metis.page.HeaderSubMenuBuilder;
 import eu.europeana.metis.templates.CssFile;
 import eu.europeana.metis.templates.Footer;
 import eu.europeana.metis.templates.FooterNavigation;
@@ -35,14 +35,15 @@ import java.util.List;
  * @author alena
  */
 public abstract class MetisPage extends AbstractMetisPage {
-  public UserDTO userDTO;
+  protected UserDTO userDTO;
 
   @Override
   public List<CssFile> resolveCssFiles() {
     CssFile cssFile1 = new CssFile();
-    CssFile cssFile2 = new CssFile();
     cssFile1.setPath("https://europeana-styleguide-test.s3.amazonaws.com/css/pandora/screen.css");
     cssFile1.setMedia("all");
+
+    CssFile cssFile2 = new CssFile();
     cssFile2.setPath("http://netdna.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.css");
     cssFile2.setMedia("all");
 
@@ -76,8 +77,7 @@ public abstract class MetisPage extends AbstractMetisPage {
   }
 
   @Override
-  public MetisHeader buildMetisHeader(PageView pageView) {
-
+  public MetisHeader buildMetisHeader() {
     Options options = new Options();
     options.setSearchActive(false);
     options.setSettingsActive(true);
@@ -92,7 +92,7 @@ public abstract class MetisPage extends AbstractMetisPage {
     menuItem.setFontawesome(true);
     menuItem.setIcon("users");
     menuItem.setIconClass("svg-icon-user-signin");
-    menuItem.setSubmenu(buildNavigationSubmenu(pageView));
+    menuItem.setSubmenu(buildNavigationSubmenu());
 
     UtilityNav utilityNav = new UtilityNav();
     utilityNav.setMenuId("settings-menu");
@@ -120,69 +120,83 @@ public abstract class MetisPage extends AbstractMetisPage {
     return metisHeader;
   }
 
-  private Submenu buildNavigationSubmenu(PageView pageView) {
-    SubmenuItem submenuItem1 = new SubmenuItem();
-    SubmenuItem submenuItem2 = new SubmenuItem();
-    submenuItem1.setIsCurrent(true);
-    submenuItem1.setIsDivider(null);
-    submenuItem1.setSubtitle(null);
-    submenuItem1.setMessage(null);
-    submenuItem1.setSubmenu(false);
-    submenuItem2.setIsCurrent(false);
-    submenuItem2.setIsDivider(null);
-    submenuItem2.setSubtitle(null);
-    submenuItem2.setMessage(null);
-    submenuItem2.setSubmenu(false);
-    switch (pageView) {
-      case REGISTER:
-        submenuItem1.setText("Login");
-        submenuItem1.setUrl("/login");
-        submenuItem2.setText("Home");
-        submenuItem2.setUrl("/");
-        break;
-      case LOGIN:
-        submenuItem1.setText("Register");
-        submenuItem1.setUrl("/register");
-        submenuItem2.setText("Home");
-        submenuItem2.setUrl("/");
-        break;
-      default:
-        if (userDTO.notNullUser() && userDTO.getLdapUser().getEmail() != null && userDTO.getLdapUser().getEmail() != null) {
-          submenuItem1.setText("Profile");
-          submenuItem1.setUrl("/profile");
-          submenuItem2.setText("Logout");
-          submenuItem2.setUrl("/logout");
-        } else {
-          submenuItem1.setText("Login");
-          submenuItem1.setUrl("/login");
-          submenuItem2.setText("Register");
-          submenuItem2.setUrl("/register");
-        }
-        break;
+  private boolean isUserAuthorized() {
+    return
+        userDTO.notNullUser() && userDTO.getLdapUser().getEmail() != null && userDTO.getLdapUser().getEmail() != null;
+  }
+
+  public Submenu buildNavigationSubmenu() {
+    if (isUserAuthorized()) {
+      return HeaderSubMenuBuilder.buildMenuWhenAuthorized();
+    } else {
+      return HeaderSubMenuBuilder.buildMenuWhenNotAuthorized();
     }
-    Submenu submenu = new Submenu();
-    submenu.setItems(Collections.newArrayList(submenuItem1, submenuItem2));
-    return submenu;
   }
 
   @Override
   public MetisFooter buildMetisFooter() {
+    MetisFooter metisFooter = new MetisFooter();
+    metisFooter.setNavigation(getFooterNavigation());
+    return metisFooter;
+  }
 
-    SubmenuItem submenuItem1 = new SubmenuItem();
-    SubmenuItem submenuItem2 = new SubmenuItem();
-    SubmenuItem submenuItem3 = new SubmenuItem();
-    SubmenuItem submenuItem4 = new SubmenuItem();
-    submenuItem1.setText("Home");
-    submenuItem1.setUrl("/");
-    submenuItem2.setText("Terms of use and policies");
-    submenuItem2.setUrl("http://europeana.eu/portal/rights/terms-and-policies.html");
-    submenuItem3.setText("Contact us");
-    submenuItem3.setUrl("/");
-    submenuItem4.setText("Sitemap");
-    submenuItem4.setUrl("/");
+  private FooterNavigation getFooterNavigation() {
+    Footer footer = new Footer();
+    footer.setSubfooter(getSubFooter());
 
+    ListOfLinks listOfLinks1 = new ListOfLinks();
+    listOfLinks1.setTitle("More Info");
+    listOfLinks1.setItems(Collections
+        .newArrayList(
+            getSubmenuItem("About", "/"),
+            getSubmenuItem("Development updates", "/"),
+            getSubmenuItem("All institutions", "/"),
+            getSubmenuItem("Become our partner", "/"),
+            getSubmenuItem("Contact us", "/")));
+
+    ListOfLinks listOfLinks2 = new ListOfLinks();
+    listOfLinks2.setTitle("Help");
+    listOfLinks2
+        .setItems(Collections.newArrayList(
+            getSubmenuItem("Search tips", "/"),
+            getSubmenuItem("Terms of Use & Policies", "/")));
+    ListOfLinks listOfLinks3 = new ListOfLinks();
+
+    listOfLinks3.setTitle("Tools");
+    listOfLinks3
+        .setItems(Collections.newArrayList(
+            getSubmenuItem("API Docs", "/"),
+            getSubmenuItem("Status", "/")));
+
+    footer.setLinklist1(listOfLinks1);
+    footer.setLinklist2(listOfLinks2);
+    footer.setLinklist3(listOfLinks3);
+    footer.setSocial(setupSocial());
+
+    FooterNavigation footerNavigation = new FooterNavigation();
+    footerNavigation.setFooter(footer);
+    return footerNavigation;
+  }
+
+  private SubFooter getSubFooter() {
     SubFooter subFooter = new SubFooter();
-    subFooter.setItems(Collections.newArrayList(submenuItem1, submenuItem2, submenuItem3, submenuItem4));
+    subFooter
+        .setItems(Collections.newArrayList(
+            getSubmenuItem("Home", "/"),
+            getSubmenuItem("Terms of use and policies", "http://europeana.eu/portal/rights/terms-and-policies.html"),
+            getSubmenuItem("Contact us", "/"),
+            getSubmenuItem("Sitemap", "/")));
+    return subFooter;
+  }
+
+  private SubmenuItem getSubmenuItem(String text, String url) {
+    SubmenuItem submenuItem = new SubmenuItem();
+    submenuItem.setText(text);
+    submenuItem.setUrl(url);
+    return submenuItem;
+  }
+
+  private Social setupSocial() {
     Social social = new Social();
     social.setFacebook(true);
     social.setGithub(false);
@@ -190,60 +204,23 @@ public abstract class MetisPage extends AbstractMetisPage {
     social.setLinkedin(false);
     social.setPinterest(true);
     social.setTwitter(true);
-
-    ListOfLinks listOfLinks1 = new ListOfLinks();
-    SubmenuItem listOflinks1SubmenuItem1 = new SubmenuItem();
-    SubmenuItem listOflinks1SubmenuItem2 = new SubmenuItem();
-    SubmenuItem listOflinks1SubmenuItem3 = new SubmenuItem();
-    SubmenuItem listOflinks1SubmenuItem4 = new SubmenuItem();
-    SubmenuItem listOflinks1SubmenuItem5 = new SubmenuItem();
-    listOflinks1SubmenuItem1.setText("About");
-    listOflinks1SubmenuItem1.setUrl("/");
-    listOflinks1SubmenuItem2.setText("Development updates");
-    listOflinks1SubmenuItem2.setUrl("/");
-    listOflinks1SubmenuItem3.setText("All institutions");
-    listOflinks1SubmenuItem3.setUrl("/");
-    listOflinks1SubmenuItem4.setText("Become our partner");
-    listOflinks1SubmenuItem4.setUrl("/");
-    listOflinks1SubmenuItem5.setText("Contact us");
-    listOflinks1SubmenuItem5.setUrl("/");
-    listOfLinks1.setTitle("More Info");
-    listOfLinks1.setItems(Collections.newArrayList(listOflinks1SubmenuItem1, listOflinks1SubmenuItem2, listOflinks1SubmenuItem3, listOflinks1SubmenuItem4, listOflinks1SubmenuItem5));
-    ListOfLinks listOfLinks2 = new ListOfLinks();
-    SubmenuItem listOflinks2SubmenuItem1 = new SubmenuItem();
-    SubmenuItem listOflinks2SubmenuItem2 = new SubmenuItem();
-    listOflinks2SubmenuItem1.setText("Search tips");
-    listOflinks2SubmenuItem1.setUrl("/");
-    listOflinks2SubmenuItem2.setText("Terms of Use & Policies");
-    listOflinks2SubmenuItem2.setUrl("/");
-    listOfLinks2.setTitle("Help");
-    listOfLinks2.setItems(Collections.newArrayList(listOflinks2SubmenuItem1, listOflinks2SubmenuItem2));
-    ListOfLinks listOfLinks3 = new ListOfLinks();
-    SubmenuItem listOflinks3SubmenuItem1 = new SubmenuItem();
-    SubmenuItem listOflinks3SubmenuItem2 = new SubmenuItem();
-    listOflinks3SubmenuItem1.setText("API Docs");
-    listOflinks3SubmenuItem1.setUrl("/");
-    listOflinks3SubmenuItem2.setText("Status");
-    listOflinks3SubmenuItem2.setUrl("/");
-    listOfLinks3.setTitle("Tools");
-    listOfLinks3.setItems(Collections.newArrayList(listOflinks3SubmenuItem1, listOflinks3SubmenuItem2));
-
-
-    Footer footer = new Footer();
-    footer.setSubfooter(subFooter);
-    footer.setLinklist1(listOfLinks1);
-    footer.setLinklist2(listOfLinks2);
-    footer.setLinklist3(listOfLinks3);
-    footer.setSocial(social);
-    FooterNavigation footerNavigation = new FooterNavigation();
-    footerNavigation.setFooter(footer);
-    MetisFooter metisFooter = new MetisFooter();
-    metisFooter.setNavigation(footerNavigation);
-
-    return metisFooter;
+    return social;
   }
 
-  public I18n buildI18n() {
+  protected I18n buildI18n() {
+    I18nGlobal i18nGlobal = new I18nGlobal();
+    i18nGlobal.setFindUsElsewhere("Find us elsewhere");
+    i18nGlobal.setMission(
+        "We transform the world with culture! We want to build on Europe’s rich heritage and make it easier for people to use, whether for work, for learning or just for fun.");
+    i18nGlobal.setMissionTitle("Our mission");
+    i18nGlobal.setNewsletter(createNewsLetter());
+    I18n i18n = new I18n();
+    i18n.setGlobal(i18nGlobal);
+
+    return i18n;
+  }
+
+  private Newsletter createNewsLetter() {
     Newsletter newsletter = new Newsletter();
     newsletter.setChooseLanguage("Choose a language");
     newsletter.setEmailAddressInvalid("Please enter a valid email address.");
@@ -251,14 +228,7 @@ public abstract class MetisPage extends AbstractMetisPage {
     newsletter.setLanguageRequired("Please choose a language for your newsletter.");
     newsletter.setSignup("Sign up for our newsletter");
     newsletter.setSubmitAlt("Subscribe");
-    I18nGlobal i18nGlobal = new I18nGlobal();
-    i18nGlobal.setFindUsElsewhere("Find us elsewhere");
-    i18nGlobal.setMission("We transform the world with culture! We want to build on Europe’s rich heritage and make it easier for people to use, whether for work, for learning or just for fun.");
-    i18nGlobal.setMissionTitle("Our mission");
-    i18nGlobal.setNewsletter(newsletter);
-    I18n i18n = new I18n();
-    i18n.setGlobal(i18nGlobal);
 
-    return i18n;
+    return newsletter;
   }
 }
