+++ metis-ui/src/main/java/eu/europeana/metis/controller/MetisProfilePageController.java
@@ -2,7 +2,7 @@ package eu.europeana.metis.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import eu.europeana.metis.page.MetisLandingPage;
-import eu.europeana.metis.page.PageView;
+import eu.europeana.metis.page.ProfileLandingPage;
 import eu.europeana.metis.ui.mongo.domain.UserDTO;
 import eu.europeana.metis.ui.mongo.service.UserService;
 import org.slf4j.Logger;
@@ -33,19 +33,23 @@ public class MetisProfilePageController {
 
   @RequestMapping(value = "/profile", method = RequestMethod.GET)
   public ModelAndView profile(Model model) throws JsonProcessingException {
+    UserDTO userDTO = getAuthenticatedUser();
+
+    ModelAndView modelAndView = new ModelAndView("templates/Pandora/Metis-Homepage");
+    MetisLandingPage metisLandingPage = new ProfileLandingPage(userDTO);
+    modelAndView.addAllObjects(metisLandingPage.buildModel());
+    return modelAndView;
+  }
+
+  private UserDTO getAuthenticatedUser() {
     Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     String primaryKey =
         principal instanceof LdapUserDetailsImpl ? ((LdapUserDetailsImpl) principal).getUsername()
             : null;
     UserDTO userDTO = userService.getUser(primaryKey);
     LOGGER.info("User profile opened: %s", userDTO.getLdapUser().getFirstName());
-
-    ModelAndView modelAndView = new ModelAndView("templates/Pandora/Metis-Homepage");
-    MetisLandingPage metisLandingPage = new MetisLandingPage(PageView.PROFILE, userDTO);
-    modelAndView.addAllObjects(metisLandingPage.buildModel());
-    return modelAndView;
+    return userDTO;
   }
-
 
 //  private List<Organization> buildAvailableOrganizationsList() {
 ////    List<String> organizations = new ArrayList<>();
