--- metis-ui/src/main/java/eu/europeana/metis/page/PageView.java
@@ -1,10 +1,0 @@
-package eu.europeana.metis.page;
-
-/**
- * 
- * @author alena
- *
- */
-public enum PageView {
-	HOME, LOGIN, REGISTER, PROFILE, REQUESTS, USER_APPROVE;
-}
